feat: add persistent sidebar navigation + fix layout and UX issues
- Add icon-only sidebar (56px) to AppComponent with auth-aware visibility;
  hidden on /login, /register, /player/* routes
- Sidebar has shield logo (links to campaigns), Campaigns icon, AI Settings
  icon, and Logout button with purple accent (#7c3aed) and Material tooltips
- Fix flex layout: exclude router-outlet from flex:1 rule so routed
  components fill the full content area (was rendering in bottom half)
- Fix combat-view height: 100vh → 100% to work inside flex shell container
- Remove redundant mat-toolbar from campaign-list, character-list, monster-list
  now that global navigation is provided by the sidebar
- Remove duplicate logout button from campaign-list toolbar
- Add Encounters button (local_fire_department icon) to campaign list actions
- Fix encounter-setup onSaveComplete() to navigate to combat view instead of /campaigns
- Add listening:false to Konva grid lines so annotation clicks pass through
- Fix index.html: add Material Icons font link and correct page title
- Backend: add CORS config, fix EncounterService BigDecimal dirty-check

+++ backend/src/main/java/com/gm/combat/service/EncounterService.java
@@ -146,7 +146,9 @@ public class EncounterService {
                     .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Monster not found"));
 
             if (monster.getChallengeRating() != null) {
-                statsOverride.put("challengeRating", monster.getChallengeRating());
+                // Store as double so JSONB round-trip (BigDecimal→JSON→Double) is consistent
+                // and Hibernate dirty-check won't trigger a spurious UPDATE
+                statsOverride.put("challengeRating", monster.getChallengeRating().doubleValue());
             }
 
             return Combatant.builder()
